[POAE7-2829] fix bugs under C++20 (#348)
* fix bugs

* update

* fix

* fix format

* fix format

* fix format

* update

* [POAE7-2848] fix bugs and clean up code (#359)

* fix codes style

---------
