+++ cpp/src/cider/exec/module/CMakeLists.txt
@@ -24,9 +24,9 @@ set(SRC_FILE
     CiderExprEvaluator.cpp
     CiderTypes.cpp
     CiderOptions.cpp
+    batch/CiderBatch.cpp
     batch/CiderBatchUtils.cpp
-    batch/CiderArrowBufferHolder.cpp
-    batch/CiderBatch.cpp)
+    batch/CiderArrowBufferHolder.cpp)
 add_library(cider SHARED ${SRC_FILE})
 target_link_libraries(
   cider
