+++ cpp/src/cider/exec/nextgen/context/CMakeLists.txt
@@ -19,6 +19,8 @@
 set(CONTEXT_SOURCE
     ${CMAKE_CURRENT_LIST_DIR}/CodegenContext.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RuntimeContext.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/Batch.cpp)
+    ${CMAKE_CURRENT_LIST_DIR}/Batch.cpp
+    ../../../exec/module/batch/CiderBatchUtils.cpp
+    ../../../exec/module/batch/CiderArrowBufferHolder.cpp)
 
 add_library(cider_context OBJECT ${CONTEXT_SOURCE})
