+++ cpp/src/cider/tests/CMakeLists.txt
@@ -106,7 +106,9 @@ if(DEFINED ENV{ISTHMUS_EXEC})
   add_subdirectory(utils)
   add_subdirectory(functionality)
   add_subdirectory(profiling)
-  add_subdirectory(benchmark)
+  if(CIDER_ENABLE_BENCHMARK)
+    add_subdirectory(benchmark)
+  endif()
   add_subdirectory(hashtable)
 endif()
