+++ cpp/src/cider/tests/CiderAggTestHelper.h
@@ -239,9 +239,6 @@ void runTest(const std::string& test_name,
              size_t buffer_entry_num = 16384,
              size_t spilled_entry_num = 0,
              const std::vector<CiderBitUtils::CiderBitVector<>>& nulls = {}) {
-  LOG(INFO) << "----------------------Test case: " + test_name +
-                   " --------------------------------------";
-
   auto cider_compile_module =
       CiderCompileModule::Make(std::make_shared<CiderDefaultAllocator>());
   auto exe_option = CiderExecutionOption::defaults();
@@ -272,11 +269,6 @@ void runTest(const std::string& test_name,
 
   CiderRuntimeModule cider_runtime_module(compile_result, compile_option, exe_option);
 
-  LOG(INFO) << "EU:\n" << *ra_exe_unit_ptr;
-  LOG(INFO) << "MemInfo\n" << cider_runtime_module.convertQueryMemDescToString();
-  LOG(INFO) << "HashTable:\n"
-            << cider_runtime_module.convertGroupByAggHashTableToString();
-
   std::unique_ptr<CiderBatch> input_batch =
       table_ptr->generateStructBatch(input_cols_name);
   for (size_t i = 0; i < input_batch->getChildrenNum(); ++i) {
@@ -298,9 +290,6 @@ void runTest(const std::string& test_name,
   for (size_t i = 0; i < 10; ++i) {
     cider_runtime_module.processNextBatch(*input_batch);
   }
-
-  LOG(INFO) << "---------------------------Execution "
-               "Success-----------------------------------";
 
   std::vector<SQLTypes> types(ra_exe_unit_ptr->target_exprs.size());
   for (size_t i = 0; i < types.size(); ++i) {
@@ -384,7 +373,5 @@ void runTest(const std::string& test_name,
     ss << "\n";
   }
 
-  LOG(INFO) << ss.str();
   CHECK_EQ(out_batch->getLength(), expect_targets_num);
-  LOG(INFO) << "Result Row Num: " << out_batch->getLength();
 }
