+++ cpp/src/cider/tests/CiderNongroupbyAggTestHelper.h
@@ -442,9 +442,6 @@ void runTest(const std::string& test_name,
              size_t buffer_entry_num = 16384,
              size_t spilled_entry_num = 0,
              const std::vector<CiderBitUtils::CiderBitVector<>>& nulls = {}) {
-  LOG(INFO) << "----------------------Test case: " + test_name +
-                   " --------------------------------------";
-
   auto cider_compile_module =
       CiderCompileModule::Make(std::make_shared<CiderDefaultAllocator>());
   auto exe_option = CiderExecutionOption::defaults();
@@ -467,11 +464,6 @@ void runTest(const std::string& test_name,
 
   CiderRuntimeModule cider_runtime_module(compile_result, compile_option, exe_option);
 
-  LOG(INFO) << "EU:\n" << *ra_exe_unit_ptr;
-  LOG(INFO) << "MemInfo\n" << cider_runtime_module.convertQueryMemDescToString();
-  LOG(INFO) << "HashTable:\n"
-            << cider_runtime_module.convertGroupByAggHashTableToString();
-
   std::unique_ptr<CiderBatch> input_batch =
       table_ptr->generateStructBatch(input_cols_name);
   for (size_t i = 0; i < input_batch->getChildrenNum(); ++i) {
@@ -491,9 +483,6 @@ void runTest(const std::string& test_name,
   }
 
   cider_runtime_module.processNextBatch(*input_batch);
-
-  LOG(INFO) << "---------------------------Execution "
-               "Success-----------------------------------";
 
   std::vector<SQLTypes> types(ra_exe_unit_ptr->target_exprs.size());
   for (size_t i = 0; i < types.size(); ++i) {
@@ -543,7 +532,6 @@ void runTest(const std::string& test_name,
     ss << "\n";
   }
 
-  LOG(INFO) << ss.str();
   verifyResult(ra_exe_unit_ptr->target_exprs[2]->get_type_info().get_type(),
                out_batch.get(),
                expect_result);
@@ -560,9 +548,6 @@ void runArrowTest(const std::string& test_name,
                   size_t buffer_entry_num = 16384,
                   size_t spilled_entry_num = 0,
                   const std::vector<CiderBitUtils::CiderBitVector<>>& nulls = {}) {
-  LOG(INFO) << "----------------------Test case: " + test_name +
-                   " --------------------------------------";
-
   auto cider_compile_module =
       CiderCompileModule::Make(std::make_shared<CiderDefaultAllocator>());
   auto exe_option = CiderExecutionOption::defaults();
@@ -580,11 +565,6 @@ void runArrowTest(const std::string& test_name,
       ra_exe_unit_ptr.get(), &table_infos, schema, compile_option, exe_option);
 
   CiderRuntimeModule cider_runtime_module(compile_result, compile_option, exe_option);
-
-  LOG(INFO) << "EU:\n" << *ra_exe_unit_ptr;
-  LOG(INFO) << "MemInfo\n" << cider_runtime_module.convertQueryMemDescToString();
-  LOG(INFO) << "HashTable:\n"
-            << cider_runtime_module.convertGroupByAggHashTableToString();
 
   std::unique_ptr<CiderBatch> input_batch =
       table_ptr->generateStructBatch(input_cols_name);
@@ -605,9 +585,6 @@ void runArrowTest(const std::string& test_name,
   }
 
   cider_runtime_module.processNextBatch(*input_batch);
-
-  LOG(INFO) << "---------------------------Execution "
-               "Success-----------------------------------";
 
   std::vector<SQLTypes> types(ra_exe_unit_ptr->target_exprs.size());
   for (size_t i = 0; i < types.size(); ++i) {
@@ -689,7 +666,6 @@ void runArrowTest(const std::string& test_name,
     ss << "\n";
   }
 
-  LOG(INFO) << ss.str();
   verifyArrowResult(ra_exe_unit_ptr->target_exprs[2]->get_type_info().get_type(),
                     out_batch.get(),
                     expect_result);
