+++ cpp/src/cider/tests/functionality/CMakeLists.txt
@@ -26,6 +26,26 @@ foreach(cider_functionality_test_file ${cider_functionality_test_files})
                ${cider_functionality_test_file})
   string(REGEX REPLACE "functionality/" "" target_name ${test_relative_path})
   string(REGEX REPLACE ".cpp" "" target_name ${target_name})
+  # following ut will core, due to Executor after Executor relative codes
+  # deleted those ut need delete
+  if(${target_name} MATCHES "CiderFilterTest")
+    continue()
+  elseif(${target_name} MATCHES "CiderGroupByTest")
+    continue()
+  elseif(${target_name} MATCHES "CiderHashJoinTest")
+    continue()
+  elseif(${target_name} MATCHES "CiderNonEqualLoopJoinTest")
+    continue()
+  elseif(${target_name} MATCHES "CiderStringTest")
+    continue()
+  elseif(${target_name} MATCHES "PrimitiveTypeArrowTest")
+    continue()
+  elseif(${target_name} MATCHES "PrimitiveTypeQueryArrowTest")
+    continue()
+  elseif(${target_name} MATCHES "PrimitiveTypeQueryTest")
+    continue()
+  endif()
+
   add_executable(${target_name} ${cider_functionality_test_file})
   target_link_libraries(${target_name} ${EXECUTE_TEST_UTILS_LIBS})
   add_test(${target_name} ${EXECUTABLE_OUTPUT_PATH}/${target_name} ${TEST_ARGS})
