+++ cpp/src/cider/tests/utils/CMakeLists.txt
@@ -36,7 +36,7 @@ set(CIDER_TESTS_UTILS_SOURCE
     CiderInt128.cpp)
 
 add_library(test_utils ${CIDER_TESTS_UTILS_SOURCE})
-target_link_libraries(test_utils cider_plan_parser cider_processor)
+target_link_libraries(test_utils duckdb cider_plan_parser cider_processor)
 
 set(TEST_ARGS "--gtest_output=xml:../")
 
@@ -87,10 +87,10 @@ target_link_libraries(CiderQueryRunnerTest ${EXECUTE_TEST_UTILS_LIBS})
 add_test(CiderQueryRunnerTest ${EXECUTABLE_OUTPUT_PATH}/CiderQueryRunnerTest
          ${TEST_ARGS})
 
-add_executable(CiderMultiBatchTest CiderMultiBatchTest.cpp)
-target_link_libraries(CiderMultiBatchTest ${EXECUTE_TEST_UTILS_LIBS})
-add_test(CiderMultiBatchTest ${EXECUTABLE_OUTPUT_PATH}/CiderMultiBatchTest
-         ${TEST_ARGS})
+# add_executable(CiderMultiBatchTest CiderMultiBatchTest.cpp)
+# target_link_libraries(CiderMultiBatchTest ${EXECUTE_TEST_UTILS_LIBS})
+# add_test(CiderMultiBatchTest ${EXECUTABLE_OUTPUT_PATH}/CiderMultiBatchTest
+# ${TEST_ARGS})
 
 add_executable(CiderRunnerTest CiderRunnerTest.cpp)
 target_link_libraries(CiderRunnerTest ${EXECUTE_TEST_UTILS_LIBS})
