+++ cpp/src/cider/tests/utils/CiderNextgenBenchmarkBase.cpp
@@ -33,7 +33,6 @@ namespace cider::test::util {
 
 void CiderNextgenBenchmarkBase::benchSQL(const std::string& sql) {
   g_enable_debug_timer = true;
-  LOG(INFO) << "BENCH SQL:" << sql;
   // duckdb
   {
     INJECT_TIMER(DuckDb);
